slack-app/requirements.txt
slack-bolt>=1.21.0
slack-sdk>=3.33.0
anthropic>=0.40.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
pytest>=8.0.0
